+++ install.sh
@@ -38,7 +38,7 @@ ALIAS_FILE="${HOME}/.aliases"
 TMP_DIR=$(mktemp -d)
 
 # Commandes disponibles
-COMMANDS=("git-ia-commit" "git-ia-commit-version" "git-ia-commit-v2" "git-ia-review" "git-ia-mr" "git-ia-mr-mcp" "git-ia-squash" "git-ia-changelog" "git-ia-explain" "git-ia-test" "git-ia-doc" "git-ia-refacto" "git-ia-menu")
+COMMANDS=("git-ia-commit" "git-ia-commit-mcp" "git-ia-commit-version" "git-ia-commit-v2" "git-ia-review" "git-ia-mr" "git-ia-mr-mcp" "git-ia-squash" "git-ia-changelog" "git-ia-explain" "git-ia-test" "git-ia-doc" "git-ia-refacto" "git-ia-menu")
 
 function cleanup {
     if [ -d "$TMP_DIR" ]; then